requirements.txt
django-simple-audit
django==1.8.6
djangorestframework
django-guardian
rest_condition
